use jest-raw-loader for importing markdown files
Allows us to import markdown examples from standalone files.

unfortunately create-react-app doesn't currently seem to allow either a custom jest.config or use of the `transform`  property of the `jest` config in `package.json`, so we're having to switch from `react-scripts test` to `jest --watch` as the test script in order to be able to use the webpack loader that allows us to import markdown files.